refactor: Move screens to screen package
