+++ noty-android/app/composeapp/src/main/java/dev/shreyaspatil/noty/composeapp/view/screen/AddNotesScreen.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package dev.shreyaspatil.noty.composeapp.view.addnotes
+package dev.shreyaspatil.noty.composeapp.view.screen
 
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
